+++ dmr/openapi/core/merger.py
@@ -1,7 +1,8 @@
 import dataclasses
 from typing import TYPE_CHECKING
 
-from dmr.openapi.objects import Components, Info, OpenAPI, Paths
+from dmr.openapi.objects import Components, Info, Paths
+from dmr.openapi.openapi import OpenAPI
 
 if TYPE_CHECKING:
     from dmr.openapi.core.context import OpenAPIContext