+++ dmr/openapi/objects/__init__.py
@@ -48,7 +48,6 @@ from dmr.openapi.objects.media_type import (
 )
 from dmr.openapi.objects.oauth_flow import OAuthFlow as OAuthFlow
 from dmr.openapi.objects.oauth_flows import OAuthFlows as OAuthFlows
-from dmr.openapi.objects.openapi import OpenAPI as OpenAPI
 from dmr.openapi.objects.operation import Operation as Operation
 from dmr.openapi.objects.parameter import Parameter as Parameter
 from dmr.openapi.objects.parameter import ParameterMetadata as ParameterMetadata